De-duplicate slash command code between edit & send composers
